+++ docs/source/user_guide/changelog.rst
@@ -34,6 +34,7 @@ Unreleased
 - Add identities to fragments when tracking single animal or single fragment videos. This allows to use idmatcher.ai afterwards.
 - Handle sessions without identification models and fragments without identities (when ``n_animals=1``) in :ref:`idmatcher.ai`.
 - Add a "*save*" button in the Segmentation App to store the computed background as an image file.
+- Do not set the automatic resolution reduction if the recommended value is too close to 1 (bigger than 0.91).
 
 6.0.12
 ======